refactor(dashboard): extract transport layer + large panels from control-client (Phase 7)
- transport/connection.ts: createMissionStream() owns the stream
  lifecycle (connect/teardown, exponential backoff 1s→30s, generation
  guard, global-stream mission filter, deferred ?mission= connect);
  control-client's stream effect now hands its event handler to the
  factory. transport/types.ts defines the stream handle/state types.
- components/: ThinkingPanel (495), MissionWorkbenchPanel (426),
  SharedFiles (485), common.tsx (110, shared helpers) extracted
  mechanically from the monolith.
- control-client.tsx: 11,823 → 10,319 lines. ToolCallItem/ChatItemRow
  family (~1.4k lines) deliberately deferred — entangled with ~20
  callback props and a dozen file-local helpers; needs the state-store
  phase first.

vitest: 221 tests pass unmodified (incl. events-reducer +
control-client.item-views); tsc --noEmit clean. No behavior change.
